+++ domain_access/src/Plugin/Action/DomainAccessEditAll.php
@@ -2,9 +2,6 @@
 
 namespace Drupal\domain_access\Plugin\Action;
 
-use Drupal\Core\Action\ActionBase;
-use Drupal\Core\Session\AccountInterface;
-
 /**
  * Assigns a user to all affiliates.
  *
@@ -14,7 +11,7 @@ use Drupal\Core\Session\AccountInterface;
  *   type = "user"
  * )
  */
-class DomainAccessEditAll extends ActionBase {
+class DomainAccessEditAll extends DomainAccessActionBase {
 
   /**
    * {@inheritdoc}
@@ -22,18 +19,6 @@ class DomainAccessEditAll extends ActionBase {
   public function execute($entity = NULL) {
     $entity->set(DOMAIN_ACCESS_ALL_FIELD, 1);
     $entity->save();
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function access($object, AccountInterface $account = NULL, $return_as_object = FALSE) {
-    // @TODO: Check this logic.
-    /** @var \Drupal\user\UserInterface $object */
-    $access = $object->access('update', $account, TRUE)
-      ->andIf($object->roles->access('edit', $account, TRUE));
-
-    return $return_as_object ? $access : $access->isAllowed();
   }
 
 }